--- a/go/otel/oteltef/anyvalue.go
+++ b/go/otel/oteltef/anyvalue.go
@@ -47,6 +47,14 @@ func (s *AnyValue) init(parentModifiedFields *modifiedFields, parentModifiedBit 
 	s.kVList.init(parentModifiedFields, parentModifiedBit)
 }
 
+func (s *AnyValue) initAlloc(parentModifiedFields *modifiedFields, parentModifiedBit uint64, allocators *Allocators) {
+	s.parentModifiedFields = parentModifiedFields
+	s.parentModifiedBit = parentModifiedBit
+
+	s.array.initAlloc(parentModifiedFields, parentModifiedBit, allocators)
+	s.kVList.initAlloc(parentModifiedFields, parentModifiedBit, allocators)
+}
+
 // reset the struct to its initial state, as if init() was just called.
 // Will not reset internal fields such as parentModifiedFields.
 func (s *AnyValue) reset() {
@@ -182,14 +190,14 @@ func (s *AnyValue) SetBytes(v pkg.Bytes) {
 	}
 }
 
-func (s *AnyValue) Clone() *AnyValue {
+func (s *AnyValue) Clone(allocators *Allocators) *AnyValue {
 	return &AnyValue{
 		string:  s.string,
 		bool:    s.bool,
 		int64:   s.int64,
 		float64: s.float64,
-		array:   s.array.Clone(),
-		kVList:  s.kVList.Clone(),
+		array:   s.array.Clone(allocators),
+		kVList:  s.kVList.Clone(allocators),
 		bytes:   s.bytes,
 	}
 }
@@ -201,6 +209,7 @@ func (s *AnyValue) byteSize() uint {
 		s.array.byteSize() + s.kVList.byteSize() + 0
 }
 
+// Copy from src to dst, overwriting existing data in dst.
 func copyAnyValue(dst *AnyValue, src *AnyValue) {
 	switch src.typ {
 	case AnyValueTypeString:
@@ -221,6 +230,30 @@ func copyAnyValue(dst *AnyValue, src *AnyValue) {
 		dst.SetBytes(src.bytes)
 	case AnyValueTypeNone:
 		dst.SetType(src.typ)
+	default:
+		panic("copyAnyValue: unexpected type: " + fmt.Sprint(src.typ))
+	}
+}
+
+// Copy from src to dst. dst is assumed to be just inited.
+func copyToNewAnyValue(dst *AnyValue, src *AnyValue, allocators *Allocators) {
+	dst.typ = src.typ
+	switch src.typ {
+	case AnyValueTypeString:
+		dst.string = src.string
+	case AnyValueTypeBool:
+		dst.bool = src.bool
+	case AnyValueTypeInt64:
+		dst.int64 = src.int64
+	case AnyValueTypeFloat64:
+		dst.float64 = src.float64
+	case AnyValueTypeArray:
+		copyToNewAnyValueArray(&dst.array, &src.array, allocators)
+	case AnyValueTypeKVList:
+		copyToNewKeyValueList(&dst.kVList, &src.kVList, allocators)
+	case AnyValueTypeBytes:
+		dst.bytes = src.bytes
+	case AnyValueTypeNone:
 	default:
 		panic("copyAnyValue: unexpected type: " + fmt.Sprint(src.typ))
 	}
@@ -682,6 +715,8 @@ type AnyValueDecoder struct {
 	isKVListRecursive bool
 
 	bytesDecoder encoders.BytesDecoder
+
+	allocators *Allocators
 }
 
 // Init is called once in the lifetime of the stream.
@@ -692,6 +727,8 @@ func (d *AnyValueDecoder) Init(state *ReaderState, columns *pkg.ReadColumnSet) e
 	}
 	state.AnyValueDecoder = d
 	defer func() { state.AnyValueDecoder = nil }()
+
+	d.allocators = &state.Allocators
 
 	var err error
 	d.fieldCount, err = state.StructFieldCounts.AnyValueFieldCount()
@@ -922,4 +959,35 @@ func (d *AnyValueDecoder) Decode(dstPtr *AnyValue) error {
 		}
 	}
 	return nil
+}
+
+// AnyValueAllocator implements a custom allocator for AnyValue.
+// It maintains a pool of pre-allocated AnyValue and grows the pool
+// dynamically as needed, up to a maximum size of 64 elements.
+type AnyValueAllocator struct {
+	pool []AnyValue
+	ofs  int
+}
+
+// Alloc returns the next available AnyValue from the pool.
+// If the pool is exhausted, it grows the pool by doubling its size
+// up to a maximum of 64 elements.
+func (a *AnyValueAllocator) Alloc() *AnyValue {
+	if a.ofs < len(a.pool) {
+		// Get the next available AnyValue from the pool
+		a.ofs++
+		return &a.pool[a.ofs-1]
+	}
+	// We've exhausted the current pool, prealloc a new pool.
+	return a.prealloc()
+}
+
+//go:noinline
+func (a *AnyValueAllocator) prealloc() *AnyValue {
+	// prealloc expands the pool by doubling its size, up to a maximum of 64 elements.
+	// If the pool is empty, it starts with 1 element.
+	newLen := min(max(len(a.pool)*2, 1), 64)
+	a.pool = make([]AnyValue, newLen)
+	a.ofs = 1
+	return &a.pool[0]
 }