--- a/go/otel/oteltef/exemplar.go
+++ b/go/otel/oteltef/exemplar.go
@@ -59,6 +59,14 @@ func (s *Exemplar) init(parentModifiedFields *modifiedFields, parentModifiedBit 
 	s.filteredAttributes.init(&s.modifiedFields, fieldModifiedExemplarFilteredAttributes)
 }
 
+func (s *Exemplar) initAlloc(parentModifiedFields *modifiedFields, parentModifiedBit uint64, allocators *Allocators) {
+	s.modifiedFields.parent = parentModifiedFields
+	s.modifiedFields.parentBit = parentModifiedBit
+
+	s.value.initAlloc(&s.modifiedFields, fieldModifiedExemplarValue, allocators)
+	s.filteredAttributes.initAlloc(&s.modifiedFields, fieldModifiedExemplarFilteredAttributes, allocators)
+}
+
 // reset the struct to its initial state, as if init() was just called.
 // Will not reset internal fields such as parentModifiedFields.
 func (s *Exemplar) reset() {
@@ -217,14 +225,17 @@ func (s *Exemplar) markUnmodifiedRecursively() {
 	s.modifiedFields.mask = 0
 }
 
-func (s *Exemplar) Clone() Exemplar {
-	return Exemplar{
+func (s *Exemplar) Clone(allocators *Allocators) Exemplar {
+
+	c := Exemplar{
+
 		timestamp:          s.timestamp,
-		value:              s.value.Clone(),
+		value:              s.value.Clone(allocators),
 		spanID:             s.spanID,
 		traceID:            s.traceID,
-		filteredAttributes: s.filteredAttributes.Clone(),
+		filteredAttributes: s.filteredAttributes.Clone(allocators),
 	}
+	return c
 }
 
 // ByteSize returns approximate memory usage in bytes. Used to calculate
@@ -234,12 +245,22 @@ func (s *Exemplar) byteSize() uint {
 		s.value.byteSize() + s.filteredAttributes.byteSize() + 0
 }
 
+// Copy from src to dst, overwriting existing data in dst.
 func copyExemplar(dst *Exemplar, src *Exemplar) {
 	dst.SetTimestamp(src.timestamp)
 	copyExemplarValue(&dst.value, &src.value)
 	dst.SetSpanID(src.spanID)
 	dst.SetTraceID(src.traceID)
 	copyAttributes(&dst.filteredAttributes, &src.filteredAttributes)
+}
+
+// Copy from src to dst. dst is assumed to be just inited.
+func copyToNewExemplar(dst *Exemplar, src *Exemplar, allocators *Allocators) {
+	dst.timestamp = src.timestamp
+	copyToNewExemplarValue(&dst.value, &src.value, allocators)
+	dst.spanID = src.spanID
+	dst.traceID = src.traceID
+	copyToNewAttributes(&dst.filteredAttributes, &src.filteredAttributes, allocators)
 }
 
 // CopyFrom() performs a deep copy from src.
@@ -396,6 +417,8 @@ type ExemplarEncoder struct {
 	filteredAttributesEncoder     *AttributesEncoder
 	isFilteredAttributesRecursive bool // Indicates FilteredAttributes field's type is recursive.
 
+	allocators *Allocators
+
 	keepFieldMask uint64
 	fieldCount    uint
 }
@@ -409,6 +432,7 @@ func (e *ExemplarEncoder) Init(state *WriterState, columns *pkg.WriteColumnSet) 
 	defer func() { state.ExemplarEncoder = nil }()
 
 	e.limiter = &state.limiter
+	e.allocators = &state.Allocators
 
 	// Number of fields in the output data schema.
 	var err error
@@ -642,6 +666,8 @@ type ExemplarDecoder struct {
 
 	filteredAttributesDecoder     *AttributesDecoder
 	isFilteredAttributesRecursive bool
+
+	allocators *Allocators
 }
 
 // Init is called once in the lifetime of the stream.
@@ -652,6 +678,8 @@ func (d *ExemplarDecoder) Init(state *ReaderState, columns *pkg.ReadColumnSet) e
 	}
 	state.ExemplarDecoder = d
 	defer func() { state.ExemplarDecoder = nil }()
+
+	d.allocators = &state.Allocators
 
 	// Number of fields in the input data schema.
 	var err error
@@ -834,4 +862,35 @@ func (d *ExemplarDecoder) Decode(dstPtr *Exemplar) error {
 	}
 
 	return nil
+}
+
+// ExemplarAllocator implements a custom allocator for Exemplar.
+// It maintains a pool of pre-allocated Exemplar and grows the pool
+// dynamically as needed, up to a maximum size of 64 elements.
+type ExemplarAllocator struct {
+	pool []Exemplar
+	ofs  int
+}
+
+// Alloc returns the next available Exemplar from the pool.
+// If the pool is exhausted, it grows the pool by doubling its size
+// up to a maximum of 64 elements.
+func (a *ExemplarAllocator) Alloc() *Exemplar {
+	if a.ofs < len(a.pool) {
+		// Get the next available Exemplar from the pool
+		a.ofs++
+		return &a.pool[a.ofs-1]
+	}
+	// We've exhausted the current pool, prealloc a new pool.
+	return a.prealloc()
+}
+
+//go:noinline
+func (a *ExemplarAllocator) prealloc() *Exemplar {
+	// prealloc expands the pool by doubling its size, up to a maximum of 64 elements.
+	// If the pool is empty, it starts with 1 element.
+	newLen := min(max(len(a.pool)*2, 1), 64)
+	a.pool = make([]Exemplar, newLen)
+	a.ofs = 1
+	return &a.pool[0]
 }