Custom allocators batch allocation requests of structs and oneofs.
The CPU improvement for array- and dict-heavy schemas is around 7-25%:

```
                                                       │ bench_main.txt │         bench_current.txt          │
                                                       │     sec/op     │   sec/op     vs base               │
Deserialization/file=deser_stef.prof/format=stef-10         698.4µ ± 1%   521.2µ ± 1%  -25.37% (p=0.000 n=9)
Deserialization/file=otelcol_otlp.prof/format=stef-10       163.5µ ± 1%   132.3µ ± 1%  -19.09% (p=0.000 n=9)
Deserialization/file=wsstreamasync.prof/format=stef-10     1052.9µ ± 0%   976.0µ ± 1%   -7.30% (p=0.000 n=9)
geomean                                                     493.6µ        406.8µ       -17.59%

                                                       │ bench_main.txt │          bench_current.txt          │
                                                       │      B/op      │     B/op      vs base               │
Deserialization/file=deser_stef.prof/format=stef-10        1.081Mi ± 0%   1.205Mi ± 0%  +11.50% (p=0.000 n=9)
Deserialization/file=otelcol_otlp.prof/format=stef-10      385.1Ki ± 0%   424.8Ki ± 0%  +10.32% (p=0.000 n=9)
Deserialization/file=wsstreamasync.prof/format=stef-10     1.516Mi ± 0%   1.619Mi ± 0%   +6.84% (p=0.000 n=9)
geomean                                                    871.3Ki        954.4Ki        +9.54%

                                                       │ bench_main.txt │         bench_current.txt          │
                                                       │   allocs/op    │  allocs/op   vs base               │
Deserialization/file=deser_stef.prof/format=stef-10        11.932k ± 0%   2.610k ± 0%  -78.13% (p=0.000 n=9)
Deserialization/file=otelcol_otlp.prof/format=stef-10       3.598k ± 0%   1.270k ± 0%  -64.70% (p=0.000 n=9)
Deserialization/file=wsstreamasync.prof/format=stef-10     15.915k ± 0%   2.253k ± 0%  -85.84% (p=0.000 n=9)
geomean                                                     8.808k        1.955k       -77.81%
```

There is an increase in memory usage, which is an acceptable tradeoff for speed
improvements and reduction in allocation counts resulting in less GC work.